Use separate file to avoid conflicts
