+++ packages/next/src/lib/verify-root-layout.ts
@@ -3,7 +3,7 @@ import * as Log from '../build/output/log'
 import { promises as fs } from 'fs'
 import { bold } from './picocolors'
 import { APP_DIR_ALIAS } from './constants'
-import type { PageExtensions } from '../build/entries'
+import type { PageExtensions } from '../build/page-extensions-type'
 
 const globOrig =
   require('next/dist/compiled/glob') as typeof import('next/dist/compiled/glob')
